Prepare v2018.01

- Move options to defconfig
- Fix sd3.0 support (by DP)
- Fix spi-nor support (by DP/Vipul)
- Add SPI_GENERIC Kconfig option (by DP/Vipul)

+++ arch/arm/cpu/armv8/generic_timer.c
@@ -43,7 +43,7 @@ unsigned long timer_read_counter(void)
 	return cntpct;
 }
 
-unsigned long long get_ticks(void)
+uint64_t get_ticks(void)
 {
 	unsigned long ticks = timer_read_counter();
